account: Clean up cancellation handling
When we have a cancellable, and it gets cancelled,
we want to call Close() on the request, to dismiss
possibly open portal dialogs. In that case, we won't
get a Response() signal, and need to complete the
GTask (with the expected G_IO_ERROR_CANCELLED error)
ourselves. Passing the cancellable to our initial
method call doesn't do any good, and only adds
confusion about who handles the cancellation, so
stop doing that.