pkg-config: Correct dependencies
We depend on gio-unix-2.0 only for linking so it is sufficient to have it in Requires.private.
gio-2.0 is required by headers, though, so it needs to be mentioned in Requires.
We probably previously omitted it because it is a required dependency of gio-unix-2.0
but now that we moved that to private dependencies, we need to mention it explicitly.
Also when we switch to Meson’s pkg-config file generator, it will do that implicitly.