OpenURI: open file with write access when requested
When user specifies to open a file with write access, we need to open it with specified
permission, otherwise xdg-desktop-portal will fail as it checks for write access.

Also follow glib and do not use O_PATH, because since xdg-desktop-portal 1.0.1, regular file descriptors
are also accepted.

Not opening the file with write access results into following error in xdg-desktop-portal:
** (/usr/libexec/xdg-desktop-portal:55877): DEBUG: 21:32:54.753: Rejecting open request for xxx
    as opening writable but fd is not writable