remote: Implement screencast restore
Add the XdpPersistMode enum, and modify the signature of
xdp_portal_create_screencast_session() to account for the
persist mode, and the restore token.

Add getters of the restore token and the persist mode to
XdpSession, and document that they are only available after
starting the session.

Adjust the corresponding code of portal tests. Make the GTK4
portal test exercise the transient persist mode.