Add API for checking for flatpak/snap sandbox
Applications need to know if they are running under a sandbox to know if
they should use portal APIs rather than direct access, in the cases that
the portal isn't built in to the toolkit and the portal isn't used by
both sandboxed and unsandboxed apps. Applications may also have other
reasons for wanting to know their sandbox status, such as Epiphany which
currently disables non-working web app functionality in that case.

Currently it's a wild west where each app has to have its own helper
function to check its sandbox status. For Flatpak one checks for the
existence of /.flatpak-info and for Snap one checks the process's
cgroups membership. The code for this (copied from xdg-desktop-portal)
is not completely trivial so it's good to have it shared between apps.
Possibly x-d-p itself should also change its libportal dependency to a
hard requirement and use the code here, but for that we would need to
allow checking by pid rather than only using the current process PID.

However adding API to check sandbox status by pid was rejected in code
review due to there not being clear users aside from xdg-desktop-portal,
and libportal being a library intended for applications not services.

One might wonder if we should have these functions use a gboolean out
parameter and reserve the return value only for the success/failure, but
that would create a lot of tedium for users of the API that call it
often or as part of complex conditional statements. Also, these
functions really shouldn't fail under any normal circumstance, but if we
add API for checking these properties of another pid other than
getpid(), then errors would be more likely (e.g. permissions).