Move the generic session declarations to a separate header
The Session is no longer unique to RemoteDesktop/ScreenCast and
shouldn't be treated as such. Let's split this out so we can use the
same object across other interfaces.

The session state is a bit more difficult since it is also related to
RD/SC only but it's more spaghettied in.