ci: add our checkout as safe for git
git rev-parse, run as part of the ABI check fails with permission errors
on current git thanks to the new security checks in git. Work around
this by marking our checkout directory as safe.

See https://github.blog/2022-04-12-git-security-vulnerability-announced/
for the details on the security issue itself.

This seems to only affect docker containers (not the github-provided
images).