tests: test Sessions.close() in the RemoteDesktop/ScreenCast templates
Previously, our templates didn't actually create sessions on the bus
because there was no need for it. To test close() on the Sesssion we
need that object to exist though.

In the respective CreateSession calls, create a Session object
on the bus and add it to a local dict. This would also allow us in the
future to check for whether sessions are valid in subsequent calls,
meanwhile it at least guarantees the DBus object path points to a
valid object.

And in the test cases intercept that call to get the libportal-assigned
session handle token and pass that to the tests. We can't get the full
session path easily (we could by extracting the first arg of the
SelectDevices/SelectSources call) but the token is good enough for the
test cases.