meson: replace the backends list with -Dbackend-foo features
This removes the 'backends' option and replaces it with individual
options for each backend, e.g. -Dbackend-gtk3=enabled.
All backends become optional by default and are built if the
dependencies are available.

A summary is printed to list the actually enabled backends.

This makes life easier for new users since the majority of them will not
have GTK3, GTK4 and Qt5 headers installed by default and thus are almost
guaranteed to run into a meson error on first configure.

Fixes https://github.com/flatpak/libportal/issues/94