Detangle Session creation from the RemoteDesktop/ScreenCast portals
This requires xdg-desktop-portal
- portal.SessionFactory, for SessionFactory.CreateSession()
- version 2 of portal.RemoteDesktop, for session_handle in RemoteDesktop.CreateSession
- version 5 of portal.ScreenCast, for session_handle in ScreenCast.CreateSession

Previously sessions were created via a session_handle_token in the
respective portal's CreateSession method. With the SessionFactory portal,
we can now create an independent Session and pass the object path
into CreateSession as session_handle. This allows us to re-use existing
sessions for multiple portals.

The existing API is updated to reflect that, in particular a session may
now have multiple types that can be checked invididually. For the
portal-specific API, the existing functions are deprecated and replaced
by one named after the portal, e.g. xdp_session_start() becomes
xdp_remote_desktop_start() or xdp_screencast_start().

The RemoteDesktop and ScreenCast portals however are a bit special,
they're quite intertwined (and partially equivalent) and the session
created for one transparently becomes the other, depending primarily on
SelectSources. This leads to some duplication in the public API.

We want new applications to use the new API, so libportal transparently
implements support for older portals by returning a special XdpSession
that is unbound but gets created on DBus when calling that portal's
CreateSession later. This session object cannot be used for future
session-based portals that require SessionPortal to be present.