remote: fix screencasting on wlroots-based compositors
xdp_portal_create_screencast_session fails if the RemoteDesktop
interface is unavailable, which is currently the case on all
wlroots-based compositors. xdp_portal_create_screencast_session queries
the version of the RemoteDesktop interface despite this interface not
beings used in the screencast session. Likewise,
xdp_portal_create_remote_desktop_session queries the version of the
ScreenCast interface despite this interface not being used in the
remote desktop session. This commit removes the unnecessary version
queries.