pyportaltest: Allow registering a session close callback
This works around a race condition where test_close_session_signal
registers a callback but on slow systems the Close signal may be sent
before said callback can be registered.

Note that we need to keep the returned SessionResult alive so the
XdpSession isn't cleaned up by the Python GC before we finish the test.

Closes https://github.com/flatpak/libportal/issues/166