notification: Filter properties based on interface version
Calls to the notification interface version 1 fail if the vardict
of `AddNotification` contains properties that aren't supported therefore
check the version and filter the vardict. This will be even more relevant
once features for version 2 are added in a future commit. The parse
function is made async because it's needed in a future commit.