notification: Ensure that g_autoptr variable is initialized
If the function early-returns (which it currently does not, but that
would be a reasonable change to make) then we'd call g_object_unref()
on uninitialized stack memory. gcc 14.2.0 in the freedesktop SDK 24.08
warns about this.

It's better to make sure that every autofreed variable is initialized,
possibly to NULL.

Signed-off-by: Simon McVittie <smcv@debian.org>