build: Never require a C++ compiler for the build architecture
Newer Meson issues a warning if we don't specify whether we plan to
compile code for the build architecture or not when adding a language.

Like most projects, this one doesn't seem to ever compile for the build
architecture, only for the host architecture, so when cross-compiling it
will only need a host-architecture C++ compiler.

For example, if we're cross-compiling for riscv64 on x86_64, we
will need a riscv64 cross-compiler, but we don't need an x86_64 native
compiler. Instruct Meson not to look for one.

Signed-off-by: Simon McVittie <smcv@debian.org>