Fix type used for signal IDs
While trying to investigate #169 I noticed that inputcapture was using
guint instead of the correct gulong for GObject signal IDs, which could
conceivably result in a signal ID being truncated and the wrong signal
being disconnected (although in practice signal IDs are allocated
starting from 0, so this is unlikely). The same pattern is seen in all
of the other portals.

Using the correct type also allows g_clear_signal_handler() to be used,
for idempotent signal disconnections.

Signed-off-by: Simon McVittie <smcv@debian.org>