tests: Run pytest with G_DEBUG=fatal-criticals
This ensures that we crash out at any sign of undefined behaviour. This
is the default for GLib's own GTest framework, but not for other test
frameworks like pytest.

Note that we cannot use fatal-warnings (which would be the default for
GTest) because at least one test-case for the inputcapture portal emulates
a faulty portal implementation, which libportal will log as a g_warning().

Signed-off-by: Simon McVittie <smcv@debian.org>