inputcapture: Reference a pre-existing session where necessary
We're passing the session object to g_task_return_pointer() with
g_object_unref as the free-function, and that's only correct if we
previously owned a reference to it. In the case where we created a fresh
session object, we did own a reference, but in the case where we merely
retrieved it from call->session, we did not.

This would have led to a use-after-free if it wasn't for the fact that
we are leaking the Call object (#190), which leaks a reference to the
session, avoiding the use-after-free via compensating errors. Before we
can safely fix the leak, we have to reference-count the session correctly.

Signed-off-by: Simon McVittie <smcv@debian.org>