inputcapture: Consistently end the Call iff the GTask has completed
After the previous commit ensured that the GTask is always given a
result at all appropriate times, this should not make any practical
difference.

However, it hopefully makes the intention clearer: after the GTask has
completed, we always dispose the Call, and if the GTask has not yet
completed, we never do.

Signed-off-by: Simon McVittie <smcv@debian.org>